% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{calculateDiscriminationForBenefit}
\alias{calculateDiscriminationForBenefit}
\title{Calculate discrimination for benefit}
\usage{
calculateDiscriminationForBenefit(data, model)
}
\arguments{
\item{data}{The validation dataset. A column named \code{trueBenefit} containing
the true treatment benefits is required}

\item{model}{The model for predicting benefit}
}
\description{
Calcualates the discrimination for benefit in a validation set based on a
model. The validation dataset is simulated and the true benefit is assumed to
be known.
}
