% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{concordanceCai}
\alias{concordanceCai}
\title{Calculate concordance}
\usage{
concordanceCai(data, limits = c(0, 0.9), quantiles = 100)
}
\arguments{
\item{data}{The evaluation data, containing columns
'predictedBenefit', 'treatment' and 'outcome'}

\item{limits}{Limit the focused area}

\item{quantiles}{The number of quantiles to be considered}
}
\description{
Calculates the concordance based on the Cai's paper (ref)
}
