% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createPredictionSettings}
\alias{createPredictionSettings}
\title{Create the prediction settings of the simulation}
\usage{
createPredictionSettings(args, fun)
}
\arguments{
\item{args}{A list containing the arguments to be passed on to the
function}

\item{fun}{The name of the function to be called for developing the
prediction model}
}
\description{
Create the settings for the development of the internal prediction model.
}
