% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{evaluateTargets}
\alias{evaluateTargets}
\title{Evaluate target settings}
\usage{
evaluateTargets(simulationSettings, analysisSettings, predictionSettings)
}
\arguments{
\item{simulationSettings}{A list containing the \code{databaseSettings},
\code{propensitySettings}, \code{baselineRiskSettings}
and the \code{treatmentEffectSettings}}

\item{analysisSettings}{...}

\item{predictionSettings}{...}
}
\description{
Evaluates the target settings for the simulation. Currently, it can be used
to derive the C-statistic of the prediction model and the event rates per
treatment arm
}
