% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSettings.R
\name{createAnalysisSettings}
\alias{createAnalysisSettings}
\title{Create the analysis settings}
\usage{
createAnalysisSettings(
  analysisId = "analysis",
  threads = 1,
  replications = 100,
  validationSize = 1e+06,
  seeds = sample(1e+05, replications),
  description = "description",
  saveDirectory = getwd()
)
}
\arguments{
\item{analysisId}{The id of the analysis.}

\item{threads}{The number of parallel threads to be considered
for running the simulations}

\item{replications}{The number of replications for runnning the
simulation}

\item{validationSize}{The size of the "true" population size. Used to
approximate the "true" performance of the method
under study}

\item{seeds}{A vector containing the seeds for each simulation
run. If not given the seeds are randomly generated}

\item{description}{The description of the simulation}

\item{saveDirectory}{The directory where the results will be stored}
}
\description{
Creates the more general settings for running the simulation.
}
